Добавлена авторизация. Обновлена документация.

+++ app/api/api_v1.py
@@ -3,13 +3,13 @@ from flask import (
     jsonify,
     abort,
     request,
-    redirect,
     url_for,
     current_app
 )
+from flask_login import current_user
 from sqlalchemy.exc import SQLAlchemyError
+from psycopg2.errors import UniqueViolation
 from marshmallow import Schema, fields
-import json
 
 from app.database import db
 
@@ -17,7 +17,12 @@ from app.database import db
 def get_args():
     args = None
     if request.form:
-        args = request.form
+        args = request.form.to_dict()
+        for k, v in args.items():
+            if v.lower() == 'true':
+                args[k] = True
+            if v.lower() == 'false':
+                args[k] = False
     elif request.json:
         args = request.json       
     return args
@@ -28,8 +33,8 @@ def convert_arg(args, arg_name, arg_type, IsNotNone=False):
     res = None
     str_type = arg_type.__name__
     msg = f'{arg_name} should be of type {str_type}'
-    if arg:
-        arg = args.get(arg_name, type=arg_type)
+    if arg is not None:
+        arg = arg_type(args.get(arg_name))
         if arg is not None:
             res = arg
         else:
@@ -57,14 +62,50 @@ def check_string(s, size, name):
 
 
 def check_arg_list(args, arg_list):
+    if len(arg_list) > 0 and args is None:
+        msg = 'There are no arguments here'
+        raise API_V1_ValidationException(msg) 
     for key in args.keys():
         if key not in arg_list:
             msg = f'The method has no argument named `{key}`'
             raise API_V1_ValidationException(msg)
 
 
+def check_authorization(only_admin):
+    if not current_user.is_authenticated:
+        raise API_V1_UnauthorizedException()
+    if only_admin and not current_user.is_admin:
+        raise API_V1_ForbiddenException()
+
+
 def log_error(msg, e):
     print(f'\n{msg}\n{e}\n')
+
+
+def try_query_func(func, resource, method):
+        try:
+            res, code = func()
+        except SQLAlchemyError as e:
+            if isinstance(e.orig, UniqueViolation):
+                msg = str(e.orig)
+                msg = msg[msg.index('Key'):-2]
+            else:
+                msg = 'Error while querying the database ' +\
+                    f'(in {resource}.{method})'
+            res = { 'message': msg }
+            code = 500
+            log_error(msg, e)
+            db.session.rollback()
+        except API_V1_ValidationException as e:
+            res = { 'message': str(e)}
+            code = 400
+        except API_V1_UnauthorizedException as e:
+            res = { 'message': str(e)}
+            code = 401
+        except API_V1_ForbiddenException as e:
+            res = { 'message': str(e)}
+            code = 403
+        return res, code
 
 
 class Error400Schema(Schema):
@@ -72,6 +113,20 @@ class Error400Schema(Schema):
         description="Сообщение об ошибке",
         required=True,
         example='Id must be should be a positive number')
+
+
+class Error401Schema(Schema):
+    message = fields.String(
+        description="Сообщение об ошибке",
+        required=True,
+        example='Unauthorized access. Make a request for /api/v1/session/login/')
+    
+
+class Error403Schema(Schema):
+    message = fields.String(
+        description="Сообщение об ошибке",
+        required=True,
+        example='Access is denied. Administrator rights are required')
 
 
 class Error404Schema(Schema):
@@ -94,6 +149,8 @@ class Error500Schema(Schema):
 api_docs = {
     'dictSchema': {
         'Error400Schema': Error400Schema,
+        'Error401Schema': Error401Schema,
+        'Error403Schema': Error403Schema,
         'Error404Schema': Error404Schema,
         'ListError404Schema': ListError404Schema,
         'Error500Schema': Error500Schema
@@ -101,12 +158,14 @@ api_docs = {
 }    
     
 class API_V1:
-    def __init__(self, name, dbclass, api_docs):
+    def __init__(self, name, dbclass, api_docs, allow_post=False, only_admin=True):
         self.bp = Blueprint(name, __name__, url_prefix=name)
         self.dbclass = dbclass
         self.name = name                        # Имя ресурса (мн. числе)
         self.tname = dbclass.__tablename__      # Имя таблицы (ед. числе)
         self.gets_param = ['limit', 'offset']   # Параметры метода GET
+        self.allow_post = allow_post            # Разрешить POST без авторизации
+        self.only_admin = only_admin            # Выполнение только с правом администратора
         
         @self.bp.route('/', methods=['GET'])
         def get():
@@ -141,30 +200,35 @@ class API_V1:
             res, code = self.delete(id)
             return jsonify(res), code
             
-
-        error_400 = f"""
-        '400':
-          description: {api_docs['errors']['400']}
+        # Формирование текста для перечисления схем ошибок + тег    
+        def errors(codes):
+            s = str(codes[0])
+            err = f"""
+        '{s}':
+          description: {api_docs['errors'][s]}
           content:
             application/json:
-              schema: Error400Schema
+              schema: Error{s}Schema
         """
-        error_404 = f"""
-        '404':
-          description: {api_docs['errors']['404']}
-          content:
-            application/json:
-              schema: Error404Schema   
-        """
-        error_500 = f"""     
-        '500':
-          description: {api_docs['errors']['500']}
-          content:
-            application/json:
-              schema: Error500Schema
+            if len(codes) > 1:
+                return err + errors(codes[1:])
+            else:
+                return err + f"""
       tags:
         - {api_docs['tags']['name']}   
-        """     
+        """
+            
+        add_codes = [400, 500]
+        edit_codes = [400, 404, 500]
+        del_codes = [404, 500]
+        if only_admin:
+            add_codes.insert(1, 403)
+            edit_codes.insert(1, 403)
+            del_codes.insert(0, 403)
+        if not allow_post:
+            add_codes.insert(1, 401)
+
+        # Установка документации к методам
         get.__doc__ = f"""
     ---
     get:
@@ -183,7 +247,7 @@ class API_V1:
           content:
             application/json:
               schema: ListError404Schema""" +\
-              error_400 + error_404 + error_500
+              errors([400, 404, 500])
         get_id.__doc__ = f"""
     ---
     get:
@@ -197,7 +261,7 @@ class API_V1:
           content:
             application/json:
               schema: {api_docs['get_id']['ouputSchema']}""" +\
-              error_400 + error_404 + error_500
+              errors([400, 404, 500])
         post.__doc__ = f"""
     ---
     post:
@@ -211,7 +275,7 @@ class API_V1:
           content:
             application/json:
               schema: {api_docs['post']['ouputSchema']}""" +\
-              error_400 + error_404 + error_500
+              errors(add_codes)
         put.__doc__ = f"""
     ---
     put:
@@ -225,7 +289,7 @@ class API_V1:
           content:
             application/json:
               schema: {api_docs['put']['ouputSchema']}""" +\
-              error_400 + error_404 + error_500
+              errors(edit_codes)
         patch.__doc__ = f"""
     ---
     patch:
@@ -239,7 +303,7 @@ class API_V1:
           content:
             application/json:
               schema: {api_docs['patch']['ouputSchema']}""" +\
-              error_400 + error_404 + error_500
+              errors(edit_codes)
         delete.__doc__ = f"""
     ---
     delete:
@@ -250,11 +314,9 @@ class API_V1:
       responses:
         '204':
           description: {api_docs['delete']['204']}""" +\
-          error_404 + error_500
+          errors(del_codes)
           
-        self.error_400 = error_400
-        self.error_404 = error_404
-        self.error_500 = error_500
+        self.errors = errors
 
     def count(self):
         try:
@@ -262,31 +324,21 @@ class API_V1:
         except SQLAlchemyError as e:
             abort(500)
         return count
-        
+
+
     def query_id(self, id, query, method):
-        try:
-            if id is None:
-                res, code = query()
+        def func():
+            data = self.dbclass.query.get(id)
+            if data is None:
+                res = { 'message':
+                    f'Resource {self.tname}.id = {id}' +\
+                    ' was not found in the database'}
+                code = 404
             else:
-                data = self.dbclass.query.get(id)
-                if data is None:
-                    res = { 'message':
-                        f'Resource {self.tname}.id = {id}' +\
-                        ' was not found in the database'}
-                    code = 404
-                else:
-                    res, code = query(id, data)            
-        except SQLAlchemyError as e:
-            msg = 'Error while querying the database' +\
-                f'(in {self.tname}.{method})'
-            res = { 'message': msg }
-            code = 500
-            log_error(msg, e)
-            db.session.rollback()
-        except API_V1_ValidationException as e:
-            res = { 'message': str(e)}
-            code = 400  
-        return res, code      
+                res, code = query(data)
+            return res, code
+        return try_query_func(func, self.tname, method)
+         
             
     def get(self, limit, offset, query=None):
         code = 200
@@ -324,13 +376,15 @@ class API_V1:
         return res, code
         
     def get_by_id(self, id, query=None):
-        def query(id, data):
+        def query(data):
             res = { 'data': data }
             return res, 200
         return self.query_id(id, query, 'GET')
     
     def post(self):
         def query():
+            if not self.allow_post:
+                check_authorization(self.only_admin)
             args = get_args()
             self.dbclass.validate_args(args)
             data = self.dbclass(**args)
@@ -339,10 +393,11 @@ class API_V1:
             db.session.commit()
             res = { 'data': data }
             return res, 201        
-        return self.query_id(None, query, 'POST')
+        return try_query_func(query, self.tname, 'POST')
         
     def put(self, id):
-        def query(id, data):
+        def query(data):
+            check_authorization(self.only_admin)
             args = get_args()
             self.dbclass.validate_args(args)
             for key in args:
@@ -353,7 +408,8 @@ class API_V1:
         return self.query_id(id, query, 'PUT')
         
     def patch(self, id):
-        def query(id, data):
+        def query(data):
+            check_authorization(self.only_admin)
             args = get_args()
             self.dbclass.validate_args(args, False)
             is_changed = False
@@ -374,7 +430,8 @@ class API_V1:
         return self.query_id(id, query, 'PATCH')
 
     def delete(self, id):
-        def query(id, data):
+        def query(data):
+            check_authorization(self.only_admin)
             db.session.delete(data)
             db.session.commit()
             return '', 204
@@ -384,6 +441,20 @@ class API_V1:
 class API_V1_ValidationException(Exception):
     def __init__(self, msg):
         self.msg = msg
+        
+    def __str__(self):
+        return self.msg
+    
+class API_V1_UnauthorizedException(Exception): 
+    def __init__(self):
+        self.msg = 'Unauthorized access. Make a request for ' + url_for('api_v1.session.login')
+        
+    def __str__(self):
+        return self.msg
+
+class API_V1_ForbiddenException(Exception):
+    def __init__(self):
+        self.msg = 'Access is denied. Administrator rights are required'
         
     def __str__(self):
         return self.msg
